+++ changelog.md
@@ -4,6 +4,7 @@
 * Added recipe to duplicate furniture with material
 * Fixed recycling recipe accepting any item
 * Added `requiredCraftingPermissionLevel` to config
+* Replaced vanilla’s ultra-slow outline renderer with a fast one
 
 # 0.1.3